Feat: App.js에 있던 component들을 각자 파일로 이동시킴
